# storage/api-docs/api-docs.json
{
  "swagger": "2.0",
  "info": {
    "title": "API Garage - Authentification JWT",
    "description": "API d'authentification pour les professionnels (JWT)",
    "version": "1.0.0",
    "contact": {
      "name": "TOOAUTO Garage"
    }
  },
  "basePath": "/api",
  "schemes": ["http", "https"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "securityDefinitions": {
    "bearerAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization",
      "description": "Token JWT (format: Bearer {token})"
    }
  },
  "security": [],
...
  "paths": {
...
    "/etablissements": {
      "get": {
        "tags": ["Etablissements"],
        "summary": "Récupérer mon établissement",
        "description": "Récupérer l'établissement du professionnel connecté",
        "operationId": "getEtablissement",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Établissement trouvé",
            "schema": { "$ref": "#/definitions/Etablissement" }
          },
          "401": { "description": "Non authentifié" },
          "404": { "description": "Aucun établissement trouvé" }
        }
      },
      "post": {
        "tags": ["Etablissements"],
        "summary": "Créer ou mettre à jour l'établissement",
        "description": "Créer un nouvel établissement ou mettre à jour celui existant. Utiliser multipart/form-data pour envoyer les fichiers logo et cover.",
        "operationId": "createOrUpdateEtablissement",
        "consumes": ["multipart/form-data"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "name", "in": "formData", "required": true, "type": "string", "description": "Nom de l'établissement" },
          { "name": "indicatif", "in": "formData", "type": "string", "default": "225" },
          { "name": "mobile", "in": "formData", "type": "string" },
          { "name": "mobile_fix", "in": "formData", "type": "string" },
          { "name": "email", "in": "formData", "type": "string" },
          { "name": "description", "in": "formData", "required": true, "type": "string" },
          { "name": "logo", "in": "formData", "type": "file", "description": "Image logo (jpeg, png, gif, webp - max 5 Mo)" },
          { "name": "cover", "in": "formData", "type": "file", "description": "Image cover (jpeg, png, gif, webp - max 5 Mo)" },
          { "name": "logo_where_is_create", "in": "formData", "type": "integer", "enum": [0, 1] },
          { "name": "cover_where_is_create", "in": "formData", "type": "integer", "enum": [0, 1] },
          { "name": "adresse", "in": "formData", "required": true, "type": "string" },
          { "name": "adresse_map", "in": "formData", "type": "string" },
          { "name": "longitude", "in": "formData", "required": true, "type": "string" },
          { "name": "latitude", "in": "formData", "required": true, "type": "string" },
          { "name": "pays_id", "in": "formData", "required": true, "type": "integer" },
          { "name": "ville_id", "in": "formData", "required": true, "type": "integer" },
          { "name": "commune_id", "in": "formData", "required": true, "type": "integer" },
          { "name": "specialite", "in": "formData", "type": "string" },
          { "name": "type_etablissement_id", "in": "formData", "required": true, "type": "integer" },
          { "name": "categorie_service_id", "in": "formData", "required": true, "type": "integer" },
          { "name": "is_whatsapp", "in": "formData", "type": "integer", "enum": [0, 1] },
          { "name": "type_de_prestations", "in": "formData", "required": true, "type": "array", "items": { "type": "integer" }, "collectionFormat": "multi", "description": "Renvoyer plusieurs fois type_de_prestations[]=2&type_de_prestations[]=3" },
          { "name": "service_mobile", "in": "formData", "type": "integer", "enum": [0, 1] },
          { "name": "cover_create_by", "in": "formData", "type": "integer", "enum": [1, 2] },
          { "name": "logo_create_by", "in": "formData", "type": "integer", "enum": [1, 2] },
          { "name": "statut", "in": "formData", "type": "integer", "enum": [0, 1] },
          { "name": "code_parrain", "in": "formData", "type": "string" },
          { "name": "is_commercial", "in": "formData", "type": "integer", "enum": [0, 1] }
        ],
        "responses": {
          "200": { "description": "Établissement mis à jour" },
          "201": { "description": "Établissement créé" },
          "401": { "description": "Non authentifié" },
          "422": { "description": "Erreur de validation" }
        }
      },
      "put": {
        "tags": ["Etablissements"],
        "summary": "Créer ou mettre à jour l'établissement (PUT)",
        "description": "Même endpoint que POST. Utiliser multipart/form-data pour les fichiers logo et cover.",
        "operationId": "createOrUpdateEtablissementPut",
        "consumes": ["multipart/form-data"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "name", "in": "formData", "required": true, "type": "string" },
          { "name": "indicatif", "in": "formData", "type": "string" },
          { "name": "mobile", "in": "formData", "type": "string" },
          { "name": "mobile_fix", "in": "formData", "type": "string" },
          { "name": "email", "in": "formData", "type": "string" },
          { "name": "description", "in": "formData", "required": true, "type": "string" },
          { "name": "logo", "in": "formData", "type": "file", "description": "Image logo - max 5 Mo" },
          { "name": "cover", "in": "formData", "type": "file", "description": "Image cover - max 5 Mo" },
          { "name": "logo_where_is_create", "in": "formData", "type": "integer", "enum": [0, 1] },
          { "name": "cover_where_is_create", "in": "formData", "type": "integer", "enum": [0, 1] },
          { "name": "adresse", "in": "formData", "required": true, "type": "string" },
          { "name": "adresse_map", "in": "formData", "type": "string" },
          { "name": "longitude", "in": "formData", "required": true, "type": "string" },
          { "name": "latitude", "in": "formData", "required": true, "type": "string" },
          { "name": "pays_id", "in": "formData", "required": true, "type": "integer" },
          { "name": "ville_id", "in": "formData", "required": true, "type": "integer" },
          { "name": "commune_id", "in": "formData", "required": true, "type": "integer" },
          { "name": "specialite", "in": "formData", "type": "string" },
          { "name": "type_etablissement_id", "in": "formData", "required": true, "type": "integer" },
          { "name": "categorie_service_id", "in": "formData", "required": true, "type": "integer" },
          { "name": "is_whatsapp", "in": "formData", "type": "integer", "enum": [0, 1] },
          { "name": "type_de_prestations", "in": "formData", "required": true, "type": "array", "items": { "type": "integer" }, "collectionFormat": "multi" },
          { "name": "service_mobile", "in": "formData", "type": "integer", "enum": [0, 1] },
          { "name": "cover_create_by", "in": "formData", "type": "integer", "enum": [1, 2] },
          { "name": "logo_create_by", "in": "formData", "type": "integer", "enum": [1, 2] },
          { "name": "statut", "in": "formData", "type": "integer", "enum": [0, 1] },
          { "name": "code_parrain", "in": "formData", "type": "string" },
          { "name": "is_commercial", "in": "formData", "type": "integer", "enum": [0, 1] }
        ],
        "responses": {
          "200": { "description": "Établissement mis à jour" },
          "201": { "description": "Établissement créé" },
          "401": { "description": "Non authentifié" },
          "422": { "description": "Erreur de validation" }
        }
      }
    },
...
  },
  "definitions": {
...
    "Etablissement": {
      "type": "object",
      "properties": {
        "id": { "type": "integer" },
        "name": { "type": "string" },
        "indicatif": { "type": "string" },
        "mobile": { "type": "string" },
        "mobile_fix": { "type": "string" },
        "email": { "type": "string" },
        "description": { "type": "string" },
        "adresse": { "type": "string" },
        "longitude": { "type": "string" },
        "latitude": { "type": "string" },
        "professionnel_id": { "type": "integer" },
        "pays_id": { "type": "integer" },
        "ville_id": { "type": "integer" },
        "commune_id": { "type": "integer" },
        "specialite": { "type": "string" },
        "type_etablissement_id": { "type": "integer" },
        "categorie_service_id": { "type": "integer" }
      }
    },
...
  }
}